content-security-policy/blink-contrib/connect-src-eventsource-blocked.sub.html

<!DOCTYPE html>
<html>

<head>
    <!-- Programmatically converted from a WebKit Reftest, please forgive resulting idiosyncrasies.-->
    <title>connect-src-eventsource-blocked</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- enforcing policy:
connect-src 'self'; script-src 'self' 'unsafe-inline';
-->
</head>

<body>
    <script>
        'use strict';

        // https://w3c.github.io/webappsec-csp/#directive-connect-src
        async_test((test) => {
            let es;

            try {
                es = new EventSource("http://www1.{{host}}:{{ports[http][0]}}/content-security-policy/blink-contrib/resources/simple-event-stream");
            } catch (e) {
                // Ch53 throws a SECURITY_ERR
                assert_equals(e.code, DOMException.SECURITY_ERR,
                    'Error must be a SECURITY_ERR.');
                test.done();
            }

            // FF48 asynchronously closes the connection.
            const poll_for_closed = () => {
                if (es.readyState === EventSource.CLOSED) {
                    test.done();
                } else {
                    requestAnimationFrame(poll_for_closed);
                }
            };

            poll_for_closed();
        });
    </script>
    <script async defer src="../support/checkReport.sub.js?reportExists=true&amp;reportField=violated-directive&amp;reportValue=connect-src%20&apos;self&apos;"></script>
</body>

</html>
